Moved Line Construction to Serialize
Previously, line construction was done by the SerializedAccount/Record
object on calling Line.  This was causing problems with trying to
cascade calls from SerializedAccount.Line to SerializedRecord.Line.

To reduce complication, Serialize() now takes the properties to output
in the Line property and the field is constructed at this time.